+++ dataiter/io.py
@@ -24,20 +24,23 @@ from dataiter import DataFrame
 from dataiter import GeoJSON
 from dataiter import ListOfDicts
 from dataiter import util
+from math import inf
 
 
-def read_csv(path, *, encoding="utf-8", sep=",", header=True, columns=[], dtypes={}):
+def read_csv(path, *, encoding="utf-8", sep=",", header=True, columns=[], strings_as_object=inf, dtypes={}):
     return DataFrame.read_csv(path,
                               encoding=encoding,
                               sep=sep,
                               header=header,
                               columns=columns,
+                              strings_as_object=strings_as_object,
                               dtypes=dtypes)
 
-def read_geojson(path, *, encoding="utf-8", columns=[], dtypes={}, **kwargs):
+def read_geojson(path, *, encoding="utf-8", columns=[], strings_as_object=inf, dtypes={}, **kwargs):
     return GeoJSON.read(path,
                         encoding=encoding,
                         columns=columns,
+                        strings_as_object=strings_as_object,
                         dtypes=dtypes,
                         **kwargs)